committing tasks

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -12,7 +12,7 @@ from random import uniform
 '''
 
 
-async def wait_random(max_delay: int = 10) -> float:
+async def wait_random(max_delay: float = 10) -> float:
     '''Generates a random number and returns it after n delay'''
     delay = uniform(0, max_delay)
     await asyncio.sleep(delay)
